backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
numpy==1.26.3
scipy==1.12.0
scikit-learn==1.4.0
sentence-transformers==2.3.1
hdbscan>=0.8.36
umap-learn==0.5.5
networkx==3.2.1
pandas==2.2.0
redis==5.0.1
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
python-dotenv==1.0.0
